+++ template-antd-vite/src/design-system/_sections/button.tsx
@@ -1,11 +1,8 @@
 import { Button, Space } from 'antd'
-import {
-  DownloadOutlined,
-  PlusOutlined,
-  SearchOutlined,
-} from '@ant-design/icons'
+import { SearchOutlined } from '@ant-design/icons'
 
 import { ComponentSection, Example } from '@/design-system/_lib/showcase'
+import { AppIcon } from '@/components/icons/icon'
 
 export default function ButtonDemo() {
   return (
@@ -38,10 +35,10 @@ export default function ButtonDemo() {
 
       <Example title="With icon">
         <Space wrap>
-          <Button type="primary" icon={<PlusOutlined />}>
+          <Button type="primary" icon={<AppIcon name="button.new" />}>
             New
           </Button>
-          <Button icon={<DownloadOutlined />}>Download</Button>
+          <Button icon={<AppIcon name="button.download" />}>Download</Button>
           <Button shape="circle" icon={<SearchOutlined />} />
         </Space>
       </Example>